Create cat-dog-years-calculator.rb
I struggled a bit with this one - it's the thing where you see an easy path involving a very long case switch statement and a lot of typing vs. a well-concealed trail to a vastly more elegant statement that scales. Took me a while to bushwhack my way to the better path. Also particularly liked the Float::INFINITY trick - first time I ever used it.